[#173] Refactor : 북마크, 댓글 도메인 해체, 레이어드 아키텍처 적용
