+++ .gitignore
@@ -41,4 +41,3 @@ application.yml
 
 ### firebase
 **/firebase/runninghi-firebase-adminsdk.json
-firebasee
