+++ src/main/java/com/runninghi/runninghibackv2/application/controller/BookmarkController.java
@@ -1,6 +1,6 @@
 package com.runninghi.runninghibackv2.application.controller;
 
-import com.runninghi.runninghibackv2.auth.jwt.JwtTokenProvider;
+import com.runninghi.runninghibackv2.common.auth.jwt.JwtTokenProvider;
 import com.runninghi.runninghibackv2.application.dto.bookmark.request.CreateBookmarkRequest;
 import com.runninghi.runninghibackv2.application.dto.bookmark.response.BookmarkedPostListResponse;
 import com.runninghi.runninghibackv2.application.dto.bookmark.response.CreateBookmarkResponse;
