+++ src/main/java/com/runninghi/runninghibackv2/postreport/application/controller/PostReportController.java
@@ -1,6 +1,6 @@
 package com.runninghi.runninghibackv2.postreport.application.controller;
 
-import com.runninghi.runninghibackv2.auth.jwt.JwtTokenProvider;
+import com.runninghi.runninghibackv2.common.auth.jwt.JwtTokenProvider;
 import com.runninghi.runninghibackv2.common.annotations.HasAccess;
 import com.runninghi.runninghibackv2.common.enumtype.ProcessingStatus;
 import com.runninghi.runninghibackv2.common.response.ApiResult;
