+++ src/test/java/com/runninghi/runninghibackv2/auto/jwt/JwtTokenProviderTests.java
@@ -1,6 +1,6 @@
 package com.runninghi.runninghibackv2.auto.jwt;
 
-import com.runninghi.runninghibackv2.auth.jwt.JwtTokenProvider;
+import com.runninghi.runninghibackv2.common.auth.jwt.JwtTokenProvider;
 import com.runninghi.runninghibackv2.common.dto.AccessTokenInfo;
 import com.runninghi.runninghibackv2.common.dto.RefreshTokenInfo;
 import com.runninghi.runninghibackv2.common.entity.Role;
